Linting output is not visible in transactional tests
