+++ crates/sui-move/src/lib.rs
@@ -15,7 +15,7 @@ pub mod build;
 pub mod coverage;
 #[cfg(feature = "disassemble")]
 pub mod disassemble;
-mod linters;
+pub mod linters;
 pub mod new;
 #[cfg(feature = "prove")]
 pub mod prove;
