+++ crates/sui-move/src/linters/mod.rs
@@ -1,4 +1,5 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod driver;
 pub mod self_transfer;
